Enable one-way pass-by-value for NSAttributedString.
one-way pass-by-value is useful for users to create NSAttributedString to feed the app, without knowing if the type is value type or not (complete value type cannot be created with GREY_REMOTE_CLASS_IN_APP).

Since it's one way, users still get by-reference NSAttributedString if it returns from an app method. Users should explicitly call -returnByValue if the actual object should be returned from the app.

PiperOrigin-RevId: 501459920