Adjust CAAnimation with 0 duration.
CAAnimation with 0 duration may still cause undefined behavior, e.g., never completes, if it doesn't set `repeatCount` and `repeatDuration` properly. This is the case when Lottie animation is set repeated (infinitely) with 0 duration.

EarlGrey by default adjusts infinite animations but the animation will bypass the adjustment in this situation. Adding extra checks for 0 duration animations to cover this case.

PiperOrigin-RevId: 536519980