Clarify EarlGrey’s requirements for a type being passed to `GREYRemoteClassInApp`.
To be compatible with EarlGrey, the class’s members need to be marked `@objc dynamic` because EarlGrey relies on Objective-C message dispatch to forward method calls across processes. The class itself does not need to be marked `@objc`, but it must inherit from `NSObject` or a subclass so that the compiler uses the Objective-C runtime for object allocation instead of the Swift runtime.

To make the above clear I made the following changes:
- Added a generic type constraint to require the target class to inherit from `NSObject`.
- Added documentation comments.

The generic type constraint also allows us to remove a force cast.

PiperOrigin-RevId: 552556313