Update ErrorHandlingTest for iOS 26.
The smaller timeout is causing failure on iOS 26 beta because some UI navigation bar animation is still running after the page is opened. Since this test is meant to test that "Paused" should not have appeared soon after the button is tapped, the timeout should be moved immediately before the assertion.

PiperOrigin-RevId: 773078276