Update AnimationsTest for iOS 26.
The failing test `testAnimationsNotPresentStringInHierarchyOnUnrelatedFailure` is due to certain animations still being included in the debugging view hierarchy. The reason seems to be that certain components in UINavigationBar have completed animations still attached (i.e. EarlGrey's synchronization is working as expected, but the printed hierarchy is confusing as it includes ended animations).

This change excludes completed animations from being included in the debugging view hierarchy.

PiperOrigin-RevId: 775838610

# Conflicts:
#	UILib/GREYElementHierarchy.m